refactor(agent): extract peer/SWM diagnostics into dkg-agent-diagnostics.ts
Cluster #1 of the dkg-agent.ts class-body split. Moves the read-only
diagnostic/stats readouts (getPeerProtocols, getSwmGossipStats,
getSwmHandlerStats, getPeerDiagnostics, pingPeers) into a new sibling
module of dependency-injection free functions, mirroring the package's
established sync/ and p2p/ extraction convention. The DKGAgent methods
remain as the public surface and delegate; the two trivial accessors
(getPeerHealth, getMessengerSloStats) stay inline.

Pure relocation, no behavior change. dkg-agent.ts -254 lines.
Typecheck clean; unit lane (123), agent.test.ts (126), and
dkg-agent-diagnostics + swm-gossip-publish-failure (28) all green.
